requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
pyrebase4==4.7.1
python-dotenv==1.0.0
Pillow==10.0.1
requests==2.31.0
Werkzeug==2.3.7